+++ packages/credential-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/credential-sdk
 
+## 0.30.0
+
+### Minor Changes
+
+- Enable auto-conversion on verification method comparison
+
 ## 0.29.0
 
 ### Minor Changes